fix(libcw): using cfg on tests
