+++ crates/libcw/Cargo.toml
@@ -12,6 +12,9 @@ categories = ["filesystem", "text-processing"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+default=[]
+
 [dependencies]
 
 [dependencies.tokio]
