+++ crates/libcw/examples/count_from_file.rs
@@ -13,6 +13,7 @@ use libcw::Parser;
 /// ```
 ///
 /// you can find the example under `target/release/examples/count_from_file`
+#[cfg(not(feature = "tokio"))]
 fn main() -> std::io::Result<()> {
     // Creates a parser with UTF8 encoding and LF linebreaks that returns
     // all stats (lines,words,chars,bytes and max length)
@@ -30,4 +31,8 @@ fn main() -> std::io::Result<()> {
     }
     println!("Took {} ms", start.elapsed().as_millis());
     Ok(())
+}
+#[cfg(feature = "tokio")]
+pub fn main() {
+    println!("Invalid feature")
 }
