+++ crates/libcw/src/parser.rs
@@ -1,5 +1,4 @@
 use std::fmt::{Display, Formatter};
-use std::io::BufRead;
 
 use crate::config::Encoding;
 use crate::config::LineBreak;
@@ -7,10 +6,14 @@ use crate::state::bytes_state::BytesState;
 use crate::state::chars_state::CharState;
 use crate::state::lines_state::LinesState;
 use crate::state::max_length::MaxLengthState;
-use crate::state::traits::{compute::Compute, partial_state::PartialState};
 use crate::state::words_state::WordsState;
 use crate::state::State;
-use crate::stats::Stats;
+
+
+#[cfg(not(feature = "tokio"))]
+pub mod features_none;
+#[cfg(feature = "tokio")]
+pub mod tokio;
 
 const BUFFER_SIZE: usize = 16 * 1024; // 8KB
 
@@ -27,7 +30,7 @@ const BUFFER_SIZE: usize = 16 * 1024; // 8KB
 ///
 /// # Example
 ///
-/// ```no_run
+/// ```ignore
 /// # use libcw::Parser;
 /// # use libcw::config::{Encoding, LineBreak};
 /// # use std::io::BufReader;
@@ -110,321 +113,4 @@ impl Parser {
             linebreak,
         }
     }
-}
-#[cfg(not(feature = "tokio"))]
-impl Parser {
-    /// The proccess method takes in a [BufRead](std::io::BufRead) instance
-    /// that is read for yielding results. If the BufRead instance cannot be
-    /// read this will yield the corresponding error
-    /// ```no_run
-    /// # use libcw::Parser;
-    /// # use libcw::config::{Encoding, LineBreak};
-    /// # use std::io::BufReader;
-    /// # use std::fs::File;
-    /// # use std::io;
-    /// # fn main() -> io::Result<()> {
-    /// let parser = Parser::new(
-    ///     Encoding::UTF8,
-    ///     LineBreak::LF,
-    ///     true,true,true,true,true
-    /// );
-    /// let read = BufReader::new(File::open("foo.txt")?);
-    /// let stats_from_read = parser.proccess(read);
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn proccess<R: BufRead + Sized>(&self, reader: R) -> std::io::Result<Stats> {
-        match self.encoding {
-            Encoding::UTF8 => self.utf8_proccess(reader),
-            Encoding::UTF16 => self.utf16_proccess(reader),
-        }
-    }
-
-    /// Runs over the tape at max speed reading utf8 encoded text
-    fn utf8_proccess<R: BufRead + Sized>(&self, mut reader: R) -> std::io::Result<Stats> {
-        let mut state = self.initial_state;
-        let mut buff = [0; BUFFER_SIZE];
-        loop {
-            let read = reader.read(&mut buff)?;
-            if read == 0 {
-                return Ok(state.output());
-            }
-            state = state.utf8_compute(&buff[0..read]);
-        }
-    }
-
-    /// Decides endianess and computes tape
-    fn utf16_proccess<R: BufRead + Sized>(&self, mut reader: R) -> std::io::Result<Stats> {
-        // TODO utf16 encoding on beta. Some test did not pass
-        let buff = reader.fill_buf()?;
-        if buff.len() < 2 {
-            // Not enought
-            let mut out = self.initial_state.output();
-            out.set_bytes(Some(buff.len()));
-            Ok(out)
-        } else {
-            let first = buff[0];
-            let second = buff[1];
-
-            if first == 0xFF && second == 0xFE {
-                // Little endian
-                let mut stats = self.initial_state.output();
-                stats.set_bytes(Some(2));
-
-                reader.consume(2);
-
-                self.utf16_process_le(reader).map(|x| x.combine(stats))
-            } else if first == 0xFE && second == 0xFF {
-                // Big endian
-                let mut stats = self.initial_state.output();
-                stats.set_bytes(Some(2));
-
-                reader.consume(2);
-
-                self.utf16_proccess_be(reader).map(|x| x.combine(stats))
-            } else {
-                // Assumed big endian
-                self.utf16_proccess_be(reader)
-            }
-        }
-    }
-    fn utf16_proccess_be<R: BufRead + Sized>(&self, mut reader: R) -> std::io::Result<Stats> {
-        let mut state = self.initial_state;
-        let mut buff = [0; BUFFER_SIZE];
-
-        let mut read = 0;
-        loop {
-            let start = if read % 2 != 0 {
-                // Put last one the first
-                buff[0] = buff[read];
-                0
-            } else {
-                // Ignore the first element
-                1
-            };
-            // [_,Some,Some,Some,Some...,BUFFER_SIZE]
-            read = reader.read(&mut buff[1..BUFFER_SIZE])?;
-
-            if read == 0 {
-                return Ok(state.output());
-            } else {
-                // Tape wont change. Non mutable call
-                state = state.utf16_compute(&buff[start..(read + 1)]);
-            }
-        }
-    }
-    fn utf16_process_le<R: BufRead + Sized>(&self, mut reader: R) -> std::io::Result<Stats> {
-        let mut state = self.initial_state;
-        let mut buff = [0; BUFFER_SIZE];
-
-        let mut read = 0;
-        loop {
-            let start = if read % 2 != 0 {
-                // Put last one the first
-                buff[0] = buff[read];
-                0
-            } else {
-                // Ignore the first element
-                1
-            };
-            // [_,Some,Some,Some,Some...,BUFFER_SIZE]
-            read = reader.read(&mut buff[1..BUFFER_SIZE])?;
-
-            for index in ((start + 1)..(read + 1)).step_by(2) {
-                buff.swap(index, index - 1)
-            }
-            /*
-            let mut index = start + 1;
-            while index < read + 1 {
-                buff.swap(index,index-1);
-                index += 2;
-            }*/
-
-            if read == 0 {
-                return Ok(state.output());
-            } else {
-                // Tape won't change. Non mutable call
-                state = state.utf16_compute(&buff[start..(read + 1)]);
-            }
-        }
-    }
-}
-#[cfg(feature = "tokio")]
-pub mod tokio {
-    use std::fmt::{Display, Formatter};
-    use std::io::BufRead;
-
-    use crate::config::Encoding;
-    use crate::config::LineBreak;
-    use crate::state::bytes_state::BytesState;
-    use crate::state::chars_state::CharState;
-    use crate::state::lines_state::LinesState;
-    use crate::state::max_length::MaxLengthState;
-    use crate::state::traits::{compute::Compute, partial_state::PartialState};
-    use crate::state::words_state::WordsState;
-    use crate::state::State;
-    use crate::stats::Stats;
-    use tokio::io::{AsyncBufRead, AsyncReadExt, AsyncBufReadExt};
-    use crate::Parser;
-    use crate::parser::BUFFER_SIZE;
-
-    impl Parser {
-        /// The proccess method takes in a [BufRead](tokio::io::BufRead) instance
-        /// that is read for yielding results. If the BufRead instance cannot be
-        /// read this will yield the corresponding error
-        /// ```no_run
-        /// # use libcw::Parser;
-        /// # use libcw::config::{Encoding, LineBreak};
-        /// use tokio::io::BufReader;
-        /// use tokio::fs::File;
-        ///
-        /// # use std::io;
-        /// # fn main() -> io::Result<()> {
-        /// let parser = Parser::new(
-        ///     Encoding::UTF8,
-        ///     LineBreak::LF,
-        ///     true,true,true,true,true
-        /// );
-        /// let read = BufReader::new(File::open("foo.txt").await?);
-        /// let stats_from_read = parser.proccess(read).await;
-        /// # Ok(())
-        /// # }
-        /// ```
-        pub async fn proccess<R>(&self, reader: R) -> std::io::Result<Stats>
-            where
-                R : AsyncReadExt + AsyncBufRead + Sized + Unpin
-        {
-            match self.encoding {
-                Encoding::UTF8 => self.utf8_proccess(reader).await,
-                Encoding::UTF16 => self.utf16_proccess(reader).await,
-            }
-        }
-
-        /// Runs over the tape at max speed reading utf8 encoded text
-        async fn utf8_proccess<R>(&self, mut reader: R) -> std::io::Result<Stats>
-            where
-                R : AsyncReadExt + AsyncBufRead + Sized + Unpin
-        {
-            let mut state = self.initial_state;
-            let mut buff = [0; BUFFER_SIZE];
-            loop {
-                let read = reader.read(&mut buff).await?;
-                if read == 0 {
-                    return Ok(state.output());
-                }
-                state = state.utf8_compute(&buff[0..read]);
-            }
-        }
-
-        /// Decides endianess and computes tape
-        async fn utf16_proccess<R>(&self, mut reader: R) -> std::io::Result<Stats>
-            where
-                R : AsyncReadExt + AsyncBufRead + Sized + Unpin
-        {
-            // TODO utf16 encoding on beta. Some test did not pass
-            let buff = reader.fill_buf().await?;
-            if buff.len() < 2 {
-                // Not enought
-                let mut out = self.initial_state.output();
-                out.set_bytes(Some(buff.len()));
-                Ok(out)
-            } else {
-                let first = buff[0];
-                let second = buff[1];
-
-                if first == 0xFF && second == 0xFE {
-                    // Little endian
-                    let mut stats = self.initial_state.output();
-                    stats.set_bytes(Some(2));
-
-                    reader.consume(2);
-
-                    self
-                        .utf16_process_le(reader)
-                        .await
-                        .map(|x| x.combine(stats))
-                } else if first == 0xFE && second == 0xFF {
-                    // Big endian
-                    let mut stats = self.initial_state.output();
-                    stats.set_bytes(Some(2));
-
-                    reader.consume(2);
-
-                    self.utf16_proccess_be(reader)
-                        .await
-                        .map(|x| x.combine(stats))
-                } else {
-                    // Assumed big endian
-                    self.utf16_proccess_be(reader).await
-                }
-            }
-        }
-        async fn utf16_proccess_be<R>(&self, mut reader: R) -> std::io::Result<Stats>
-            where
-                R : AsyncReadExt + AsyncBufRead + Sized + Unpin
-        {
-            let mut state = self.initial_state;
-            let mut buff = [0; BUFFER_SIZE];
-
-            let mut read = 0;
-            loop {
-                let start = if read % 2 != 0 {
-                    // Put last one the first
-                    buff[0] = buff[read];
-                    0
-                } else {
-                    // Ignore the first element
-                    1
-                };
-                // [_,Some,Some,Some,Some...,BUFFER_SIZE]
-                read = reader.read(&mut buff[1..BUFFER_SIZE]).await?;
-
-                if read == 0 {
-                    return Ok(state.output());
-                } else {
-                    // Tape wont change. Non mutable call
-                    state = state.utf16_compute(&buff[start..(read + 1)]);
-                }
-            }
-        }
-        async fn utf16_process_le<R>(&self, mut reader: R) -> std::io::Result<Stats>
-            where
-                R : AsyncReadExt + AsyncBufRead + Sized + Unpin
-        {
-            let mut state = self.initial_state;
-            let mut buff = [0; BUFFER_SIZE];
-
-            let mut read = 0;
-            loop {
-                let start = if read % 2 != 0 {
-                    // Put last one the first
-                    buff[0] = buff[read];
-                    0
-                } else {
-                    // Ignore the first element
-                    1
-                };
-                // [_,Some,Some,Some,Some...,BUFFER_SIZE]
-                read = reader.read(&mut buff[1..BUFFER_SIZE]).await?;
-
-                for index in ((start + 1)..(read + 1)).step_by(2) {
-                    buff.swap(index, index - 1)
-                }
-                /*
-                let mut index = start + 1;
-                while index < read + 1 {
-                    buff.swap(index,index-1);
-                    index += 2;
-                }*/
-
-                if read == 0 {
-                    return Ok(state.output());
-                } else {
-                    // Tape won't change. Non mutable call
-                    state = state.utf16_compute(&buff[start..(read + 1)]);
-                }
-            }
-        }
-    }
-
 }
